+++ src/main/java/com/hotsix/iAmNotAlone/signup/domain/Region.java
@@ -1,15 +1,10 @@
-package com.hotsix.iAmNotAlone.domain;
+package com.hotsix.iAmNotAlone.signup.domain;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import java.util.ArrayList;
-import java.util.List;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -29,9 +24,5 @@ public class Region {
     private Long id;
     private String sido;
     private String sigg;
-
-//    @JsonManagedReference
-//    @OneToMany(mappedBy = "region")
-//    private List<Member> members = new ArrayList<>();
 
 }
