+++ src/main/java/com/hotsix/iAmNotAlone/signup/domain/dto/MemberDto.java
@@ -1,8 +1,8 @@
-package com.hotsix.iAmNotAlone.domain.dto;
+package com.hotsix.iAmNotAlone.signup.domain.dto;
 
-import com.hotsix.iAmNotAlone.domain.Member;
-import com.hotsix.iAmNotAlone.domain.Personality;
-import com.hotsix.iAmNotAlone.domain.Region;
+import com.hotsix.iAmNotAlone.signup.domain.Member;
+import com.hotsix.iAmNotAlone.signup.domain.Personality;
+import com.hotsix.iAmNotAlone.signup.domain.Region;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.Builder;
@@ -18,7 +18,6 @@ public class MemberDto {
     private LocalDate birth;
     private String introduction;
     private int gender;
-    private Long regionId;
     private Region region;
     private List<Personality> personalities;
 
@@ -30,9 +29,9 @@ public class MemberDto {
             .birth(member.getBirth())
             .introduction(member.getIntroduction())
             .gender(member.getGender())
-            .regionId(member.getRegion().getId())
             .region(member.getRegion())
             .personalities(member.getPersonalities())
             .build();
     }
+
 }
