+++ src/main/java/com/hotsix/iAmNotAlone/signup/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.hotsix.iAmNotAlone.repository;
+package com.hotsix.iAmNotAlone.signup.repository;
 
-import com.hotsix.iAmNotAlone.domain.Member;
+import com.hotsix.iAmNotAlone.signup.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
